pyproject.toml
[project]
name = "mason-ai"
version = "0.1.0"
description = "An opinionated toolkit for small language models customization"
readme = "README.md"
authors = [
    { name = "Alejandro Piad", email = "apiad@apiad.net" }
]
requires-python = ">=3.13"
dependencies = [
    "llama-cpp-python>=0.3.16",
    "peft>=0.18.0",
    "rich>=14.2.0",
    "transformers>=4.57.1",
    "typer>=0.20.0",
]

[tool.hatch.build.targets.wheel]
packages = ["mason"]

[project.scripts]
mason = "mason.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
